+++ app/build.gradle
@@ -1,6 +1,5 @@
 apply plugin: 'com.android.application'
 
-
 android {
     compileSdkVersion 27
     buildToolsVersion "27.0.3"
@@ -22,7 +21,6 @@ android {
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support.constraint:constraint-layout:1.0.2'
     androidTestImplementation('com.android.support.test.espresso:espresso-core:2.2.2', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
@@ -30,8 +28,5 @@ dependencies {
     implementation 'com.android.support:design:27.1.1'
     // Google Play Services dependency
     implementation 'com.google.android.gms:play-services-auth:12.0.1'
-    implementation 'com.google.firebase:firebase-core:12.0.1'
-    implementation 'com.google.firebase:firebase-auth:12.0.1'
     testImplementation 'junit:junit:4.12'
 }
-//apply plugin: 'com.google.gms.google-services'
