+++ app/src/main/java/com/regent/tech/med_manager/WelcomeActivity.java
@@ -1,75 +1,13 @@
 package com.regent.tech.med_manager;
 
-import android.content.Intent;
-import android.support.annotation.NonNull;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.util.Log;
-import android.view.Menu;
-import android.view.MenuItem;
-import android.view.View;
-import android.widget.Button;
 
-import com.google.android.gms.auth.api.signin.GoogleSignIn;
-import com.google.android.gms.auth.api.signin.GoogleSignInClient;
-import com.google.android.gms.auth.api.signin.GoogleSignInOptions;
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.Task;
-
-public class WelcomeActivity extends AppCompatActivity{
-
-    private GoogleSignInClient mGoogleSignInClient;
-
+public class WelcomeActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_welcome);
-
-        GoogleSignInOptions gso = new GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
-                .requestEmail()
-                .build();
-        mGoogleSignInClient = GoogleSignIn.getClient(this, gso);
     }
-
-    // [START signOut]
-    public void signOut() {
-        mGoogleSignInClient.signOut()
-                .addOnCompleteListener(this, new OnCompleteListener<Void>() {
-                    @Override
-                    public void onComplete(@NonNull Task<Void> task) {
-                        // [START_EXCLUDE]
-                        Intent intent = new Intent(WelcomeActivity.this, LoginActivity.class);
-                        startActivity(intent);
-                        // [END_EXCLUDE]
-                    }
-                });
-    }
-    // [END signOut]
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu){
-        // Inflate the menu options from the res/menu/menu_login.xml file.
-        // This adds menu items to the app bar.
-        getMenuInflater().inflate(R.menu.menu_login, menu);
-        return true;
-    }
-
-
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        // User clicked on a menu option in the app bar overflow menu
-        switch (item.getItemId()) {
-            // Respond to a click on the "Sign out" menu option
-            case R.id.sign_out:
-                signOut();
-                return true;
-            case R.id.disconnect:
-//                loginActivity.revokeAccess();
-                return true;
-        }
-        return super.onOptionsItemSelected(item);
-    }
-
 }
