+++ app/src/main/res/layout/activity_welcome.xml
@@ -6,14 +6,6 @@
     android:layout_height="match_parent"
     tools:context="com.regent.tech.med_manager.WelcomeActivity">
 
-    <Button
-        android:id="@+id/test_sign_out"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/sign_out"
-        android:layout_alignParentTop="true"
-        android:layout_alignParentRight="true"/>
-
     <ImageView
         android:id="@+id/profile_picture"
         android:layout_width="wrap_content"
