+++ app/src/main/res/menu/menu_login.xml
@@ -3,7 +3,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="WelcomeActivity">
+    tools:context=".LoginActivity">
 
     <item
         android:id="@+id/sign_out"
