+++ app/src/main/res/values/strings.xml
@@ -18,7 +18,5 @@
     <!-- TODO(user): replace with your real server client ID -->
     <!-- Server Client ID.  This should be a valid Web OAuth 2.0 Client ID obtained
          from https://console.developers.google.com/ -->
-    <!--<string name="server_client_id">948863978074-84o50os5kqhfln5l0oqgfbt1a4ot8bv4.apps.googleusercontent.com</string>-->
-    <!--<string name="server_client_id">1054952682027-mg26va51c991s4gpf2u3ldbqk5bif4qp.apps.googleusercontent.com</string>-->
-    <string name="server_client_id">1054952682027-11fraa65ll0917cro1kg56rrosomlg4s.apps.googleusercontent.com</string>
+    <string name="server_client_id">948863978074-84o50os5kqhfln5l0oqgfbt1a4ot8bv4.apps.googleusercontent.com</string>
 </resources>
